// src/theme/default-theme/colors/types.ts
export interface DefaultPrimaryType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultSecondaryType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultBlackType {
  0: string;
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  850: string;
  900: string;
  A100: string;
  A200: string;
  A300: string;
}

export interface DefaultTransparentType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultTransparentInverseType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultAppType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultSuccessType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultImportantType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultWarningType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultInformationType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultCautionType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultPendingType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultRiskType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultBuyType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultSellType {
  100: string;
  200: string;
  300: string;
  400: string;
  50: string;
  500: string;
  600: string;
  700: string;
  800: string;
  900: string;
}

export interface DefaultStatusType {
  accepted: string;
  acceptedText: string;
  amended: string;
  amendedText: string;
  balanceCancelled: string;
  balanceCancelledText: string;
  buy: string;
  calculated: string;
  calculatedText: string;
  cancelled: string;
  cancelledText: string;
  closed: string;
  closedText: string;
  completed: string;
  completedText: string;
  corrected: string;
  correctedText: string;
  doneForDay: string;
  doneForDayText: string;
  expired: string;
  expiredText: string;
  filled: string;
  filledText: string;
  halted: string;
  haltedText: string;
  new: string;
  newText: string;
  open: string;
  openText: string;
  orderStatus: string;
  orderStatusText: string;
  partCancelled: string;
  partCancelledText: string;
  partFilled: string;
  partFilledText: string;
  pending: string;
  pendingCancel: string;
  pendingCancelText: string;
  pendingNew: string;
  pendingNewText: string;
  pendingText: string;
  preClose: string;
  preCloseText: string;
  preOpen: string;
  preOpenText: string;
  rejected: string;
  rejectedText: string;
  replaced: string;
  replacedText: string;
  requestRejected: string;
  requestRejectedText: string;
  restatement: string;
  restatementText: string;
  sell: string;
  sending: string;
  sendingText: string;
  trade: string;
  tradeCancel: string;
  tradeCancelText: string;
  tradeText: string;
  unknown: string;
  unknownText: string;
  working: string;
  workingText: string;
}

export interface DefaultShadowsType {
  0: string;
  1: string;
  10: string;
  11: string;
  12: string;
  13: string;
  14: string;
  15: string;
  16: string;
  17: string;
  18: string;
  19: string;
  2: string;
  20: string;
  21: string;
  22: string;
  23: string;
  24: string;
  3: string;
  4: string;
  5: string;
  6: string;
  7: string;
  8: string;
  9: string;
  L1: string;
  L10: string;
  L11: string;
  L12: string;
  L13: string;
  L14: string;
  L15: string;
  L16: string;
  L17: string;
  L18: string;
  L19: string;
  L2: string;
  L20: string;
  L21: string;
  L22: string;
  L23: string;
  L24: string;
  L3: string;
  L4: string;
  L5: string;
  L6: string;
  L7: string;
  L8: string;
  L9: string;
}